+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix warning in web regarding missing emoji font subset
 - Center the bar custom drinks empty state card
+- Removing a photo from a custom drink now also deletes the image from DB
 
 ## [1.1.0] - 2026-04-10
 ### Added